Start improving comments
There are many comments in the code, but also much space for improvement.
Here, we start improving comments by:
* Translating all the Russian comments into English;
* Rewriting certain comments to make them more clear;
* Removing some redundant comments.

This work has to be continued.

Here we also add the implementation of the Monotone partitioning algorithm
for polygon triangulation which has been written about 4 years ago but
hasn't been added to the repository at that time.